+++ EduNLP/Vector/QuesNet/__init__.py
@@ -1,0 +1,1 @@
+from .QuesNet import QuesNetModel
