+++ EduNLP/Vector/__init__.py
@@ -7,3 +7,4 @@ from .rnn import RNNModel
 from .t2v import T2V, get_pretrained_t2v, PRETRAINED_MODELS
 from .embedding import Embedding
 from .bert_vec import BertModel
+from .QuesNet import QuesNetModel
