+++ EduNLP/Vector/t2v.py
@@ -7,6 +7,7 @@ from EduData import get_data
 from .rnn import RNNModel
 from .gensim_vec import W2V, D2V
 from .bert_vec import BertModel
+from .QuesNet import QuesNetModel
 from .meta import Vector
 from EduNLP.constant import MODEL_DIR
 
@@ -18,7 +19,8 @@ MODELS = {
     "lstm": RNNModel,
     "gru": RNNModel,
     "elmo": RNNModel,
-    'bert': BertModel
+    'bert': BertModel,
+    'quesnet': QuesNetModel
 }
